requirements/dev.txt
-r base.txt

pytest-django
pytest-sugar